frontend/requirements.txt
kivy==2.2.0
kivymd==1.1.1
requests==2.31.0
websocket-client==1.6.4
python-dotenv==1.0.0
pillow==10.1.0
qrcode[pil]==7.4.2
cloudinary==1.36.0

